+++ src/lightning/pytorch/CHANGELOG.md
@@ -271,6 +271,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 -
 
 
+- Fixed early stopping triggering extra validation runs after reaching `min_epochs` or `min_steps` ([#16719](https://github.com/Lightning-AI/lightning/pull/16719))
+
+
 ## [1.9.1] - 2023-02-10
 
 ### Fixed